+++ src/main/org/firebirdsql/jdbc/AbstractResultSet.java
@@ -1617,7 +1617,7 @@ public abstract class AbstractResultSet implements ResultSet, FirebirdResultSet,
     }
 
     private FBBlob asFBBlob(Blob blob) throws SQLException {
-        // if the passed BLOB is not instance of our class, copy its content into the our BLOB
+        // if the passed BLOB is not instance of our class, copy its content into our BLOB
         if (blob == null) {
             return null;
         }
@@ -1666,7 +1666,7 @@ public abstract class AbstractResultSet implements ResultSet, FirebirdResultSet,
     }
 
     private FBClob asFBClob(Clob clob) throws SQLException {
-        // if the passed BLOB is not instance of our class, copy its content into the our BLOB
+        // if the passed BLOB is not instance of our class, copy its content into our CLOB
         if (clob == null) {
             return null;
         }